use llama binary instead of extending ruby
Using the binary simplifies the code, but the biggest reason for this
change is that wrapping the binary is significantly faster. I'm not
sure why this is, but the same C++ code compiled as a Ruby extension
takes about 4x as long to run.

I'm keeping the code in ext/ because it's sort of like a native
extension in that it compiles code on the user's machine. This seems
like the least confusing way to manage that. (E.g. if the compilation
fails, the user gets an error saying "ERROR: Failed to build gem native
extension" rather than "some random script failed.")